Started on creation of enemy automization
So, here is the plan:
IN the preliniary stages, enemies are going to be soft-coded with a
StateActionTrain, which will describe the action to be taken in case of
______. This will require the addition of a recursive search in PFNC,
and possibly the addition of something to allow for enemies to select
between potential targets.
Also, another thing that needs adding is a feature by which enemies can
neglect a target and start firing upon an enemy who becomes more
obvious.